+++ staging/src/k8s.io/apiserver/pkg/storage/etcd3/testing/test_server.go
@@ -46,5 +46,8 @@ func NewUnsecuredEtcd3TestClientServer(t *testing.T) (*EtcdTestServer, *storageb
 		},
 		Paging: true,
 	}
+	if err := config.Transport.Complete(); err != nil {
+		t.Fatal(err)
+	}
 	return server, config
 }
